+++ lib/online_migrations/schema_statements.rb
@@ -701,7 +701,7 @@ module OnlineMigrations
     # @see https://edgeapi.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/SchemaStatements.html#method-i-add_foreign_key
     #
     def add_foreign_key(from_table, to_table, validate: true, **options)
-      if foreign_key_exists?(from_table, **options.merge(to_table: to_table))
+      if foreign_key_exists?(from_table, to_table, **options)
         message = "Foreign key was not created because it already exists " \
           "(this can be due to an aborted migration or similar): from_table: #{from_table}, to_table: #{to_table}".dup
         message << ", #{options.inspect}" if options.any?